Task progress is now with a progress bar instead of images, for better handing varying number of players

+++ game/src/main/java/com/team2/game/GamingController/RegisterService.java
@@ -247,7 +247,7 @@ public class RegisterService {
         }
     }
 
-    public int taskResolved(String gameID, String sessionId, String task){
+    public float taskResolved(String gameID, String sessionId, String task){
         return groupManager.getGameInstance(gameID).taskResolved(sessionId, task);
     }
